HDPI-1429: Move some event pages into new package
