+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/page/resumepossessionclaim/ClaimTypeNotEligibleWales.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.pcs.ccd.page.createpossessionclaim;
+package uk.gov.hmcts.reform.pcs.ccd.page.resumepossessionclaim;
 
 import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
 import uk.gov.hmcts.ccd.sdk.api.callback.AboutToStartOrSubmitResponse;
@@ -11,34 +11,23 @@ import java.util.List;
 
 import static uk.gov.hmcts.reform.pcs.ccd.ShowConditions.NEVER_SHOW;
 
-public class ClaimantTypeNotEligibleEngland implements CcdPageConfiguration {
+public class ClaimTypeNotEligibleWales implements CcdPageConfiguration {
 
     @Override
     public void addTo(PageBuilder pageBuilder) {
         pageBuilder
-            .page("claimantTypeNotEligibleEngland", this::midEvent)
+            .page("claimTypeNotEligibleWales", this::midEvent)
             .pageLabel("You're not eligible for this online service")
-            .showCondition("showClaimantTypeNotEligibleEngland=\"Yes\"")
-            .readonly(PCSCase::getShowClaimantTypeNotEligibleEngland, NEVER_SHOW)
-            .label("claimantTypeNotEligibleEngland-info", """
+            .showCondition("showClaimTypeNotEligibleWales=\"Yes\"")
+            .readonly(PCSCase::getShowClaimTypeNotEligibleWales, NEVER_SHOW)
+            .label("claimTypeNotEligibleWales-info", """
                 ---
                 <p class="govuk-body">
-                This service is currently only available for registered providers of social housing.
+                You cannot make a trespass claim using this service.
 
                 <h3 class="govuk-heading-s govuk-!-font-size-19">What to do next</h3>
 
-                <ul class="govuk-list">
-                  <li class="govuk-!-font-size-19">
-                    <span class="govuk-!-font-weight-bold">For rental or mortgage arrears claims</span> – use the
-                    <a href="https://www.gov.uk/possession-claim-online-recover-property"
-                        target="_blank"
-                        class="govuk-link">Possession Claim Online (PCOL) service (opens in a new tab)</a>.
-                  </li>
-                  <li class="govuk-!-font-size-19">
-                    <span class="govuk-!-font-weight-bold">For other types of claims</span> – use form N5
-                    and the correct particulars of claim form.
-                  </li>
-                </ul>
+                Use form N5 Wales and the correct particulars of claim form.
 
                 <a href="https://www.gov.uk/government/collections/property-possession-forms" rel="noreferrer noopener"
                     target="_blank"
