+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/page/resumepossessionclaim/GroundsForPossession.java
@@ -1,11 +1,11 @@
-package uk.gov.hmcts.reform.pcs.ccd.page.createpossessionclaim;
+package uk.gov.hmcts.reform.pcs.ccd.page.resumepossessionclaim;
 
 import uk.gov.hmcts.reform.pcs.ccd.common.CcdPageConfiguration;
 import uk.gov.hmcts.reform.pcs.ccd.common.PageBuilder;
 import uk.gov.hmcts.reform.pcs.ccd.domain.PCSCase;
 
 /**
- * Placeholder page configuration for the Grounds for Possession section. 
+ * Placeholder page configuration for the Grounds for Possession section.
  * Full implementation will be done in another ticket - responses not captured at the moment.
  */
 public class GroundsForPossession implements CcdPageConfiguration {
