+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/page/resumepossessionclaim/ResumeClaim.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.pcs.ccd.page.createpossessionclaim;
+package uk.gov.hmcts.reform.pcs.ccd.page.resumepossessionclaim;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
