+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/page/resumepossessionclaim/TenancyLicenceDetails.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.pcs.ccd.page.createpossessionclaim;
+package uk.gov.hmcts.reform.pcs.ccd.page.resumepossessionclaim;
 
 
 import org.springframework.beans.factory.annotation.Qualifier;
